moving some things around
