--- packages/newspringchurchapp/src/prayer/PrayerMenu/PrayerMenu.stories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react-native';
-import { PaddedView } from '@apollosproject/ui-kit';
-
-import PrayerMenu from '.';
-
-storiesOf('PrayerMenu', module).add('Example', () => (
-  <PaddedView>
-    <PrayerMenu />
-  </PaddedView>
-));
